fix: C++17-safe matrix_bit_cast (restore GPU/CPU pipeline build) + GPU-1 verified
Colab T4 run: cell 4c (test_gpu_agg.cu) GREEN — GPU SUM/MIN/MAX/COUNT == matrix_cpu_reduce
exactly (incl. >2^32 SUM 549755289600 and empty sentinels). GPU-1 (AGG-2) is hardware-
verified; the cross-backend invariant holds.

But cells 4 & 5 (full pipeline, nvcc/g++ -std=c++17) failed: std::bit_cast is C++20, and
those cells use c++17 (run_tests.sh builds main at c++20, so it was never caught locally).
Fixed by replacing std::bit_cast with a memcpy-based matrix_bit_cast (compute.hpp) at the
5 compute_mock.cpp sites — restores the proven c++17 nvcc/g++ pipeline config. Verified:
clang++ -std=c++17 main.cpp compiles + oracle 83886070; 60-test c++20 suite still green.
Also cleaned the two cosmetic warnings the run surfaced (unused-result, printf %llu).

+++ test_migration_gpu.cpp
@@ -45,6 +45,7 @@ int main() {
     assert(col.checksum() == want && "checksum invariant HOST->DEVICE->HOST");
 
     std::printf("PASS: VRAM-promoted column is GPU-scannable and byte-intact "
-                "(gpu_count=%llu cpu_count=%llu)\n", gpu_count, cpu_count);
+                "(gpu_count=%llu cpu_count=%llu)\n",
+                static_cast<unsigned long long>(gpu_count), static_cast<unsigned long long>(cpu_count));
     return 0;
 }
